+++ demo/DemoV3/Model/LongRunningfunctionStatus.generated.cs
@@ -18,13 +18,14 @@ namespace DemoV3.Model {
         public virtual string Status {
             get => _status;
             set {
-                if (!StatusValues.AllowedValues.Contains(value))
+                var _value = StatusValues.AllowedValues.FirstOrDefault(v=>String.Compare(v, value, StringComparison.OrdinalIgnoreCase) == 0, value);
+                if (!StatusValues.AllowedValues.Contains(_value))
                 {
                     var allowedValues = string.Join(", ", StatusValues.AllowedValues.Select(v => v.ToString()));
                     throw new ArgumentOutOfRangeException("status",
                         $"{value} is not allowed. Allowed values: [{allowedValues}]");
                 }
-                _status = value;
+                _status = _value;
             }
         }