+++ changelog.md
@@ -1,5 +1,8 @@
 # caffoa changelog
 
+### 1.7.0
+* Allow to ignore the case of string enums
+
 ### 1.6.0
 * added time parser for `type: string format: time` that uses DateTimeOnly or Timespan to represent times in the format HH:mm:ss. Parser also accepts h:m.
 * do not fail on body different from application/json. Instead, warn an generate endpoint with Stream